Create BibTeX entries from a package

Extract the information of a package to BibTeX. This is done by creating a cff object with cff_create() and extracting the corresponding entries with cff_to_bibtex().

Usage

cff_extract_to_bibtex(x, what = "preferred")

Arguments

x

The source that would be used for generating the cff object. It could be:

A missing value. That would retrieve the DESCRIPTION file on your in-development package.

An existing cff object,

The name of an installed package ("jsonlite"), or

Path to a DESCRIPTION file ("*/DESCRIPTION*").

what

Fields to extract. The value could be:

preferred: This would create a single entry with the main citation info of the package.

references: Extract all the entries on references.

all: A combination of the previous two options. This would extract both the preferred citation info and the references.

Value

A bibentry object or a list of bibentry objects. This could be parsed to BibTeX using toBibtex()

See also

Examples

# \donttest{

jsonvalidate <- cff_extract_to_bibtex("jsonvalidate")
#> Warning: no date field in DESCRIPTION file of package ‘jsonvalidate’

jsonvalidate
#> FitzJohn R, Ashton R, Buus M, Poberezkin E (2022). _jsonvalidate:
#> Validate 'JSON' Schema_. https://docs.ropensci.org/jsonvalidate/,
#> https://github.com/ropensci/jsonvalidate.

toBibtex(jsonvalidate)
#> @Manual{fitzjohnashton:2022,
#>   title = {jsonvalidate: Validate 'JSON' Schema},
#>   author = {Rich FitzJohn and Rob Ashton and Mathias Buus and Evgeny Poberezkin},
#>   year = {2022},
#>   note = {https://docs.ropensci.org/jsonvalidate/, https://github.com/ropensci/jsonvalidate},
#> }

lite <- cff_extract_to_bibtex("jsonlite", "references")

lite
#> R Core Team (2022). “methods.” Depends, <https://www.R-project.org/>.
#> 
#> Wickham H (2022). “httr.” Suggests.
#> 
#> Ooms J (2022). “curl.” Suggests.
#> 
#> Wickham H, Henry L, Vaughan D (2022). “vctrs.” Suggests,
#> <https://vctrs.r-lib.org/>.
#> 
#> Wickham H (2022). “testthat.” Suggests.
#> 
#> Xie Y (2022). “knitr.” Suggests, <https://yihui.org/knitr/>.
#> 
#> Allaire J, Xie Y, McPherson J, Luraschi J, Ushey K, Atkins A, Wickham
#> H, Cheng J, Chang W, Iannone R (2022). “rmarkdown.” Suggests.

toBibtex(lite)
#> @Misc{rcoreteam:2022,
#>   title = {methods},
#>   author = {{R Core Team}},
#>   year = {2022},
#>   address = {Vienna, Austria},
#>   url = {https://www.R-project.org/},
#>   note = {Depends},
#>   institution = {R Foundation for Statistical Computing},
#>   abstract = {R: A Language and Environment for Statistical Computing},
#> }
#> 
#> @Misc{wickham:2022,
#>   title = {httr},
#>   author = {Hadley Wickham},
#>   year = {2022},
#>   note = {Suggests},
#>   abstract = {httr: Tools for Working with URLs and HTTP},
#> }
#> 
#> @Misc{ooms:2022,
#>   title = {curl},
#>   author = {Jeroen Ooms},
#>   year = {2022},
#>   note = {Suggests},
#>   abstract = {curl: A Modern and Flexible Web Client for R},
#> }
#> 
#> @Misc{wickhamhenry:2022,
#>   title = {vctrs},
#>   author = {Hadley Wickham and Lionel Henry and Davis Vaughan},
#>   year = {2022},
#>   url = {https://vctrs.r-lib.org/},
#>   note = {Suggests},
#>   abstract = {vctrs: Vector Helpers},
#> }
#> 
#> @Misc{wickham:2022,
#>   title = {testthat},
#>   author = {Hadley Wickham},
#>   year = {2022},
#>   note = {Suggests},
#>   abstract = {testthat: Unit Testing for R},
#> }
#> 
#> @Misc{xie:2022,
#>   title = {knitr},
#>   author = {Yihui Xie},
#>   year = {2022},
#>   url = {https://yihui.org/knitr/},
#>   note = {Suggests},
#>   abstract = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
#> }
#> 
#> @Misc{allairexie:2022,
#>   title = {rmarkdown},
#>   author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
#>   year = {2022},
#>   note = {Suggests},
#>   abstract = {rmarkdown: Dynamic Documents for R},
#> }
# }