Parse a person to cff

Parse a person or string to a valid format for a CITATION.cff file.

cff_parse_person(person)

Arguments

person
A person object or a character coercible to person. See utils::person().

Value

A cff object ready to be used on cff_create().

Details

This is a helper function designed to help on adding or replacing the auto-generated authors of the package. See Examples.

See also

Examples

# Parse a person object

cff_parse_person(person(
  given = "First",
  family = "Author",
  role = c("aut", "cre"),
  email = "first.last@example.com",
  comment = c(
    ORCID = "0000-0001-8457-4658",
    affiliation = "An affiliation"
  )
))

#> family-names: Author
#> given-names: First
#> email: first.last@example.com
#> orcid: https://orcid.org/0000-0001-8457-4658
#> affiliation: An affiliation

# Parse a string

cff_parse_person("Julio Iglesias <fake@email.com>")

#> family-names: Iglesias
#> given-names: Julio
#> email: fake@email.com

# Several persons
persons <- c(person("Clark", "Kent"), person("Lois", "Lane"))

cff_parse_person(persons)

#> - family-names: Kent
#>   given-names: Clark
#> - family-names: Lane
#>   given-names: Lois