Create a BibTeX entry from a CITATION file or a cff object

Creates a bibentry object (bibentry()) from a cff object

Usage

cff_to_bibtex(x)

Arguments

x

The source that would be used for generating the cff object. It could be:

An existing cff object,

A CITATION.cff file.

Value

A bibentry object that can be parsed to BibTeX format with toBibtex()

References

Patashnik, Oren. "BIBTEXTING" February 1988. https://osl.ugr.es/CTAN/biblio/bibtex/base/btxdoc.pdf.

Haines, R., & The Ruby Citation File Format Developers. (2021). Ruby CFF Library (Version 0.9.0) (Computer software). doi:10.5281/zenodo.1184077 .

See also

Examples

# \donttest{

# From a cff object
package <- cff_create("rmarkdown")

obj <- cff_to_bibtex(package)

obj
#> Allaire J, Xie Y, McPherson J, Luraschi J, Ushey K, Atkins A, Wickham
#> H, Cheng J, Chang W, Iannone R (2022). _rmarkdown: Dynamic Documents
#> for R_. R package version 2.14, <https://github.com/rstudio/rmarkdown>.

toBibtex(obj)
#> @Manual{allairexie:2022,
#>   title = {rmarkdown: Dynamic Documents for R},
#>   author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
#>   year = {2022},
#>   url = {https://github.com/rstudio/rmarkdown},
#>   note = {R package version 2.14},
#> }
# }