Use a git pre-commit hook

Install a pre-commit hook that remembers you to update your CITATION.cff file. This is a wrapper of usethis::use_git_hook().

Usage

cff_git_hook_install()

cff_git_hook_remove()

Value

Invisible. This function is called for its side effects.

Details

This function would install a pre-commit hook using usethis::use_git_hook().

A pre-commit hook is a script that identifies simple issues before submission to code review. This pre-commit hook would warn you if any of the following conditions are met:

You included in a commit your DESCRIPTION or inst/CITATION file, you are not including your CITATION.cff and the CITATION.cff file is "older" than any of your DESCRIPTION or inst/CITATION file.

You have updated your CITATION.cff but you are not including it on your commit.

A word of caution

The pre-commit hook may prevent you to commit if you are not updating your CITATION.cff. However, the mechanism of detection is not perfect and would be triggered also even if you have tried to update your CITATION.cff file.

This is typically the case when you have updated your DESCRIPTION or inst/CITATION files but those changes doesn't make a change on your CITATION.cff file (i.e. you are including new dependencies).

In those cases, you can override the check running git commit --no-verify on the terminal.

If you are using RStudio you can run also this command from a R script by selecting that line and sending it to the terminal using:

Windows & Linux: Ctrl+Alt+Enter.

Mac: Cmd+Option+Return.

Removing the git pre-commit hook

You can remove the pre-commit hook by running cff_git_hook_remove().

See also

Examples

if (FALSE) {
cff_git_hook_install()
}