Read BibTeX markup as a cff_ref_lst object

Convert one or more complete BibTeX entries supplied as a character vector into a cff_ref_lst object.

Usage

cff_read_bib_text(x, encoding = "UTF-8", ...)

Arguments

x

A character vector with one or more complete BibTeX entries.

encoding

Encoding to be assumed for x. See base::readLines().

...

Arguments passed to cff_read_bib().

Value

An object of classes cff_ref_lst, cff according to the definitions.reference specified in the following guide: Citation File Format schema guide. Each element of the cff_ref_lst object has classes cff_ref, cff.

Details

This function writes x to a temporary *.bib file and reads it using cff_read_bib().

This function requires bibtex (>= 0.5.0) and uses bibtex::read.bib() for parsing.

See also

Examples

if (requireNamespace("bibtex", quietly = TRUE)) {
  x <- c(
    "@book{einstein1921,
      title        = {Relativity: The Special and the General Theory},
      author       = {Einstein, Albert},
      year         = 1920,
      publisher    = {Henry Holt and Company},
      address      = {London, United Kingdom},
      isbn         = 9781587340925
  }",
    "@misc{misc-full,
      title        = {Handing out random pamphlets in airports},
      author       = {Joe-Bob Missilany},
      year         = 1984,
      month        = oct,
      note         = {This is a full MISC entry},
      howpublished = {Handed out at O'Hare}
  }"
  )

  cff_read_bib_text(x)
}
#> - type: book
#>   title: 'Relativity: The Special and the General Theory'
#>   authors:
#>   - family-names: Einstein
#>     given-names: Albert
#>   year: '1920'
#>   publisher:
#>     name: Henry Holt and Company
#>     address: London, United Kingdom
#>   isbn: '9781587340925'
#> - type: generic
#>   title: Handing out random pamphlets in airports
#>   authors:
#>   - family-names: Missilany
#>     given-names: Joe-Bob
#>   year: '1984'
#>   month: '10'
#>   notes: This is a full MISC entry
#>   medium: Handed out at O'Hare