Schema utils

Helper functions with the valid values of different fields, according to the Citation File Format schema version 1.2.0.

cff_schema_keys() provides the valid high-level keys of the Citation File Format.

cff_schema_keys_license() provides the valid SPDX license identifier(s) to be used on the CITATION.cff file.

cff_schema_definitions_person() and cff_schema_definitions_entity() returns the valid fields to be included when defining a person or entity.

cff_schema_definitions_refs() provides the valid keys to be used on the preferred-citation and references keys.

cff_schema_keys(sorted = FALSE)

cff_schema_keys_license()

cff_schema_definitions_person()

cff_schema_definitions_entity()

cff_schema_definitions_refs()

Arguments

sorted
Logical TRUE/FALSE. Should the keys be arranged alphabetically?

Source

Guide to Citation File Format schema version 1.2.0.

Value

A vector of characters with the names of the valid keys to be used on a Citation File Format version 1.2.0

Examples

cff_schema_keys(sorted = TRUE)

#>  [1] "abstract"            "authors"             "cff-version"        
#>  [4] "commit"              "contact"             "date-released"      
#>  [7] "doi"                 "identifiers"         "keywords"           
#> [10] "license"             "license-url"         "message"            
#> [13] "preferred-citation"  "references"          "repository"         
#> [16] "repository-artifact" "repository-code"     "title"              
#> [19] "type"                "url"                 "version"            

# Valid Licenses keys
head(cff_schema_keys_license(), 20)

#>  [1] "0BSD"              "AAL"               "Abstyles"         
#>  [4] "Adobe-2006"        "Adobe-Glyph"       "ADSL"             
#>  [7] "AFL-1.1"           "AFL-1.2"           "AFL-2.0"          
#> [10] "AFL-2.1"           "AFL-3.0"           "Afmparse"         
#> [13] "AGPL-1.0-only"     "AGPL-1.0-or-later" "AGPL-3.0-only"    
#> [16] "AGPL-3.0-or-later" "Aladdin"           "AMDPLPA"          
#> [19] "AML"               "AMPAS"            

cff_schema_definitions_person()

#>  [1] "address"       "affiliation"   "alias"         "city"         
#>  [5] "country"       "email"         "family-names"  "fax"          
#>  [9] "given-names"   "name-particle" "name-suffix"   "orcid"        
#> [13] "post-code"     "region"        "tel"           "website"      

cff_schema_definitions_entity()

#>  [1] "address"    "alias"      "city"       "country"    "date-end"  
#>  [6] "date-start" "email"      "fax"        "location"   "name"      
#> [11] "orcid"      "post-code"  "region"     "tel"        "website"   

cff_schema_definitions_refs()

#>  [1] "abbreviation"        "abstract"            "authors"            
#>  [4] "collection-doi"      "collection-title"    "collection-type"    
#>  [7] "commit"              "conference"          "contact"            
#> [10] "copyright"           "data-type"           "database-provider"  
#> [13] "database"            "date-accessed"       "date-downloaded"    
#> [16] "date-published"      "date-released"       "department"         
#> [19] "doi"                 "edition"             "editors"            
#> [22] "editors-series"      "end"                 "entry"              
#> [25] "filename"            "format"              "identifiers"        
#> [28] "institution"         "isbn"                "issn"               
#> [31] "issue"               "issue-date"          "issue-title"        
#> [34] "journal"             "keywords"            "languages"          
#> [37] "license"             "license-url"         "loc-end"            
#> [40] "loc-start"           "location"            "medium"             
#> [43] "month"               "nihmsid"             "notes"              
#> [46] "number"              "number-volumes"      "patent-states"      
#> [49] "pmcid"               "publisher"           "recipients"         
#> [52] "repository"          "repository-artifact" "repository-code"    
#> [55] "scope"               "section"             "senders"            
#> [58] "status"              "term"                "thesis-type"        
#> [61] "title"               "translators"         "type"               
#> [64] "url"                 "version"             "volume"             
#> [67] "volume-title"        "year"                "year-original"