Validate a CITATION.cff file or a cff object

Validate a CITATION.cff file or a cff object created with cff_create() using the corresponding validation schema.json.

cff_validate(x = "./CITATION.cff")

Arguments

x
This is expected to be either a cff object created with cff_create() or the path to a CITATION.cff file to be validated.

Value

A message indicating the result of the validation and an invisible value TRUE/FALSE.

See also

Examples

# Full .cff example
cff_validate(system.file("examples/CITATION_complete.cff", package = "cffr"))

#> 
#> cff_validate results-----

#> Congratulations! This .cff file is valid

# Validate a cffr object
cffr <- cff_create("jsonlite")
class(cffr)

#> [1] "cff"

cff_validate(cffr)

#> 
#> cff_validate results-----

#> Congratulations! This cff object is valid

# .cff with errors
cff_validate(system.file("examples/CITATION_error.cff", package = "cffr"))

#> 
#> cff_validate results-----

#> Oops! This  .cff file has the following errors:
#> 
#> 

#>             field                          message
#> 1            data        has additional properties
#> 2  data.authors.0                 no schemas match
#> 3        data.doi referenced schema does not match
#> 4 data.keywords.0                is the wrong type
#> 5    data.license referenced schema does not match
#> 6        data.url referenced schema does not match

if (FALSE) {
# If a CITATION file (note that is not .cff) it throws an error
cff_validate(system.file("CITATION", package = "cffr"))
}