cffr: Generate Citation File Format ('CFF') Metadata for R Packages

Citation File Format ('CFF') version 1.2.0 doi:10.5281/zenodo.5171937 is a human- and machine-readable file format for software citation metadata. Core utilities generate, read, write and validate Citation File Format metadata for R packages.

References

Hernangómez D (2021). "cffr: Generate Citation File Format Metadata for R Packages." Journal of Open Source Software, 6(67), 3900. doi:10.21105/joss.03900 . https://joss.theoj.org/papers/10.21105/joss.03900.

Druskat S, Spaaks JH, Chue Hong N, Haines R, Baker J, Bliven S, Willighagen E, Pérez-Suárez D, Konovalov A (2021). "Citation File Format." doi:10.5281/zenodo.5171937 . https://github.com/citation-file-format/citation-file-format.

See also

Author

Maintainer: Diego Hernangómez diego.hernangomezherrero@gmail.com (ORCID) [copyright holder]

Authors:

Diego Hernangómez diego.hernangomezherrero@gmail.com (ORCID) [copyright holder]

Other contributors:

João Martins (ORCID) [reviewer]

Scott Chamberlain (ORCID) [reviewer]