Previous API: create BibTeX entries from multiple sources

Use as_bibentry() instead.

Usage

cff_extract_to_bibtex(x, what = c("preferred", "references", "all"))

cff_to_bibtex(x, what = c("preferred", "references", "all"))

Arguments

x

The source used to generate the bibentry object with cffr. It can be:

A missing or NULL value, which retrieves the DESCRIPTION file from your in-development package.

An existing cff object created with cff(), cff_create() or as_cff().

A path to a CITATION.cff file ("CITATION.cff").

The name of an installed package ("jsonlite").

A path to a DESCRIPTION file ("DESCRIPTION").

what

Fields to extract from a full cff object. It can be:

preferred: Create a single entry with the main citation information of the package (key preferred-citation).

references: Extract all entries of the references key.

all: Extract both the preferred-citation and references keys.

See vignette("r-cff", package = "cffr").

Value

See as_bibentry().

See also

Examples

# \donttest{
# From a `cff` object.
cff_object <- cff()

cff_object
#> cff-version: 1.2.0
#> message: If you use this software, please cite it using these metadata.
#> title: My Research Software
#> authors:
#> - family-names: Doe
#>   given-names: John

# A `bibentry` object.
bib <- as_bibentry(cff_object)
bib
#> Doe J (????). “My Research Software.”
# }